OS-7228. Fixed migration for premigrated clusters
## Description

Fixed migration for premigrated clusters

## Related issue number

OS-7228

## Checklist

* [ ] The pull request title is a good summary of the changes
* [ ] Unit tests for the changes exist
* [ ] New and existing unit tests pass locally

+++ diworker/diworker/migrations/2024012914311500_combine_aws_indexes.py
@@ -47,7 +47,7 @@ class Migration(BaseMigration):
                 partialFilterExpression=partial_exp)
             LOG.info('Added index %s', new_index_name)
         for index_name in old_indexes_map.keys():
-            if index_name not in new_indexes_map:
+            if index_name not in new_indexes_map and index_name in existing_index_map:
                 self.raw_collection.drop_index(index_name)
                 LOG.info('Dropped index %s', index_name)
